working on link from homeview for userapp
