+++ App.js
@@ -1,22 +1,19 @@
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import { StyleSheet, Text, View, Button } from 'react-native';
+import { createStackNavigator } from 'react-navigation'
+import ReportPhoto from './components/ReportPhoto.js'
+import HomeScreen from './components/HomeScreen.js'
+
+const NavLinks = createStackNavigator(
+  {
+    Home: HomeScreen,
+    ReportPhoto: ReportPhoto
+  },
+  { initialRouteName: 'Home' }
+)
 
 export default class App extends React.Component {
   render() {
-    return (
-      <View style={styles.container}>
-        <Text>Pothole Patrol</Text>
-        <Text>Let's make Chicago better, together.</Text>
-      </View>
-    );
+    return <NavLinks />
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
