+++ package.json
@@ -21,6 +21,8 @@
   "dependencies": {
     "expo": "^27.0.1",
     "react": "16.3.1",
-    "react-native": "~0.55.2"
+    "react-native": "~0.55.2",
+    "react-navigation": "^2.4.1",
+    "watchman": "^1.0.0"
   }
 }
